requirements.txt
numpy==1.24.3
pandas==2.0.3
scikit-learn==1.3.0
matplotlib==3.7.2
seaborn==0.12.2
jupyter==1.0.0
notebook==7.0.2
ipykernel==6.25.0
django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
requests==2.31.0
